New subdomain and orgName for Sh:erpa
Hi,
There are some updates in the names and URLs. The org on Github is now called SherpaCLI and the Netlify subdomain sherpa-cli.

Moved from Astro to NuxtJS for the new website. 

Out of curiosity... Any idea of the scheduled published date of the updated 'examples' page? The live one have "Last updated: 09/11/2024" in the footer.